project/tests/test_model.py
import torch
import pytest
from project.model import MyAwesomeModel

def test_model():
    print('test my awesome model')
    model = MyAwesomeModel(0.3)
    x = torch.randn(1, 1, 28, 28)
    y = model(x)
    assert y.shape == (1, 10)

@pytest.mark.parametrize("batch_size", [32, 64])
def test_model(batch_size: int) -> None:
    model = MyAwesomeModel(0.3)
    x = torch.randn(batch_size, 1, 28, 28)
    y = model(x)
    assert y.shape == (batch_size, 10)
