project/requirements.txt
torch==2.5.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
invoke==2.2.0
matplotlib==3.10.0  
hydra-core==1.3.2 
torchvision==0.20.1